requirements.txt
django
django-heroku
psycopg2
djangorestframework
